switch cleanly between clover and default
